# backend/src/main/resources/openapi/switchboard-api.yaml
openapi: 3.0.3
info:
  title: Switchboard API
  description: AI-powered feature flag platform. Management surface (user bearer auth), SDK surface (SDK-key auth), and job triggers.
  version: 0.1.0
servers:
  - url: http://localhost:28080
security:
  - bearerAuth: []
...
paths:
...
  /api/orgs:
    post:
      tags: [Orgs]
      operationId: createOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgCreateRequest'
      responses:
        '201':
          description: Created org
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgResponse'
    get:
      tags: [Orgs]
      operationId: listOrgs
      responses:
        '200':
          description: Orgs the caller belongs to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgResponse'
...
  /api/orgs/{orgId}/members:
    get:
      tags: [Orgs]
      operationId: listOrgMembers
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgMemberResponse'
    post:
      tags: [Orgs]
      operationId: addOrgMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgMemberAddRequest'
      responses:
        '201':
          description: Added member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMemberResponse'
  /api/orgs/{orgId}/members/{userId}:
    delete:
      tags: [Orgs]
      operationId: removeOrgMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed
  /api/orgs/{orgId}/settings:
    get:
      tags: [Orgs]
      operationId: getOrgSettings
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Org settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgSettingsResponse'
    put:
      tags: [Orgs]
      operationId: updateOrgSettings
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgSettingsUpdateRequest'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgSettingsResponse'
...
  /api/eval:
    post:
      tags: [Evaluation]
      operationId: evaluateAll
      security:
        - sdkKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkEvalRequest'
      responses:
        '200':
          description: Evaluation results for every non-archived flag in the key's environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkEvalResponse'
  /api/eval/{flagKey}:
    post:
      tags: [Evaluation]
      operationId: evaluateFlag
      security:
        - sdkKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/FlagKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleEvalRequest'
      responses:
        '200':
          description: Evaluation result. Unknown flag echoes the supplied default with reason SDK_DEFAULT - never an error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalResult'
  /api/eval/bootstrap:
    get:
      tags: [Evaluation]
      operationId: getBootstrap
      security:
        - sdkKeyAuth: []
      parameters:
        - name: contextKey
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Full environment flag payload; ETag is the environment stateVersion.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '304':
          description: Not modified (If-None-Match matched current stateVersion)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Firebase ID token, or dev token (Bearer dev:email) in local profile.
    sdkKeyAuth:
      type: http
      scheme: bearer
      description: Environment SDK key (sb_srv_...).
    ofrepApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        OFREP's apiKey scheme, carrying the same environment SDK key (sb_srv_...) as sdkKeyAuth.
        Both headers resolve to the same SdkKeyPrincipal; Authorization wins when both are sent.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FlagKey:
      name: flagKey
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    OrgRole:
      type: string
      enum: [OWNER, MEMBER]
    OrgCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    OrgResponse:
      type: object
      required: [id, name, slug, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        role:
          $ref: '#/components/schemas/OrgRole'
        createdAt:
          type: string
          format: date-time
    OrgMemberResponse:
      type: object
      required: [userId, email, role, joinedAt]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/OrgRole'
        joinedAt:
          type: string
          format: date-time
    OrgMemberAddRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          $ref: '#/components/schemas/OrgRole'
    OrgSettingsResponse:
      type: object
      required: [aiEnabled, autoRollbackEnabled, autoOptimizeEnabled, staleFlagWeeks]
      properties:
        aiEnabled:
          type: boolean
        autoRollbackEnabled:
          type: boolean
        autoOptimizeEnabled:
          type: boolean
        staleFlagWeeks:
          type: integer
        notificationWebhookSet:
          type: boolean
    OrgSettingsUpdateRequest:
      type: object
      properties:
        aiEnabled:
          type: boolean
        autoRollbackEnabled:
          type: boolean
        autoOptimizeEnabled:
          type: boolean
        staleFlagWeeks:
          type: integer
          minimum: 1
          maximum: 52
        notificationWebhookUrl:
          type: string
...
    FlagKind:
      type: string
      enum: [BOOLEAN, STRING]
    Variation:
      type: object
      required: [id, value]
      properties:
        id:
          type: string
          format: uuid
        value:
          type: string
        name:
          type: string
...
    ClauseOp:
      type: string
      enum: [EQUALS, IN, CONTAINS, STARTS_WITH, SEGMENT_MATCH, NOT_SEGMENT_MATCH]
    Clause:
      type: object
      required: [attribute, op, values]
      properties:
        attribute:
          type: string
          description: Context attribute name; 'key' targets the context key. For SEGMENT_MATCH ops, values carry segment keys.
        op:
          $ref: '#/components/schemas/ClauseOp'
        values:
          type: array
          items:
            type: string
    WeightedVariation:
      type: object
      required: [variationId, weight]
      properties:
        variationId:
          type: string
          format: uuid
        weight:
          type: integer
          minimum: 0
          maximum: 100
    RolloutOrVariation:
      type: object
      description: Exactly one of variationId or rollout must be set.
      properties:
        variationId:
          type: string
          format: uuid
        rollout:
          type: array
          description: Weights must sum to exactly 100.
          items:
            $ref: '#/components/schemas/WeightedVariation'
    IndividualTarget:
      type: object
      required: [contextKey, variationId]
      properties:
        contextKey:
          type: string
        variationId:
          type: string
          format: uuid
    Rule:
      type: object
      required: [id, clauses, serve]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        clauses:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Clause'
        serve:
          $ref: '#/components/schemas/RolloutOrVariation'
    FlagTargetingConfig:
      type: object
      required: [offVariationId, defaultVariationId, fallthrough]
      properties:
        individualTargets:
          type: array
          items:
            $ref: '#/components/schemas/IndividualTarget'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
        fallthrough:
          $ref: '#/components/schemas/RolloutOrVariation'
        offVariationId:
          type: string
          format: uuid
        defaultVariationId:
          type: string
          format: uuid
...
    SegmentRule:
      type: object
      required: [clauses]
      properties:
        clauses:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Clause'
...
    EvalContext:
      type: object
      required: [key]
      properties:
        key:
          type: string
          description: Stable context key (e.g. user id). Bucketing input.
        attributes:
          type: object
          additionalProperties:
            type: string
    BulkEvalRequest:
      type: object
      required: [context]
      properties:
        context:
          $ref: '#/components/schemas/EvalContext'
    SingleEvalRequest:
      type: object
      required: [context]
      properties:
        context:
          $ref: '#/components/schemas/EvalContext'
        default:
          type: string
          description: Value echoed back (reason SDK_DEFAULT) when the flag is unknown.
    EvalReason:
      type: string
      enum: [KILL_SWITCH, FLAG_OFF, TARGET_MATCH, RULE_MATCH, ROLLOUT, DEFAULT, SDK_DEFAULT]
    EvalResult:
      type: object
      required: [flagKey, value, reason]
      properties:
        flagKey:
          type: string
        variationId:
          type: string
          format: uuid
        value:
          type: string
          description: Variation value; boolean flags serve "true"/"false".
        reason:
          $ref: '#/components/schemas/EvalReason'
        ruleId:
          type: string
          format: uuid
        flagVersion:
          type: integer
    BulkEvalResponse:
      type: object
      required: [stateVersion, results]
      properties:
        stateVersion:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/EvalResult'
    BootstrapFlag:
      type: object
      required: [key, kind, variations, enabled, killSwitchActive, config, version]
      properties:
        key:
          type: string
        kind:
          $ref: '#/components/schemas/FlagKind'
        variations:
          type: array
          items:
            $ref: '#/components/schemas/Variation'
        enabled:
          type: boolean
        killSwitchActive:
          type: boolean
        config:
          $ref: '#/components/schemas/FlagTargetingConfig'
        version:
          type: integer
    BootstrapSegment:
      type: object
      required: [key, includedKeys, excludedKeys, rules]
      properties:
        key:
          type: string
        includedKeys:
          type: array
          items:
            type: string
        excludedKeys:
          type: array
          items:
            type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/SegmentRule'
    BootstrapResponse:
      type: object
      required: [envKey, stateVersion, flags, segments]
      properties:
        envKey:
          type: string
        stateVersion:
          type: integer
          format: int64
        flags:
          type: array
          items:
            $ref: '#/components/schemas/BootstrapFlag'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/BootstrapSegment'
...